Remove support for farmOS v1
