+++ farmOS/client.py
@@ -1,146 +1,84 @@
 import logging
-from urllib.parse import parse_qs, urlparse
+from urllib.parse import urlparse
 
 logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
 
 
-class BaseAPI:
-    """Base class for API methods
+class ResourceBase:
+    """Base class for JSONAPI resource methods."""
 
-    This class includes basic implementations of the
-    get(), send() and delete() methods used in the API client.
-
-    Keyword Arguments:
-    session: APISession for the farmOS instance
-    entity_type: String, used to set the entity type in the path
-        of all requests used with the API
-    """
-
-    def __init__(self, session, entity_type=None):
+    def __init__(self, session):
         self.session = session
-        self.entity_type = entity_type
-        self.filters = {}
+        self.params = {}
 
-    def _get_single_record_data(self, id):
-        """Retrieve one record given the record ID"""
-        # Set path to return record type by specific ID
-        path = self.entity_type + "/" + str(id) + ".json"
+    def _get_records(self, entity_type, bundle=None, resource_id=None, params=None):
+        """Helper function that checks to retrieve one record, one page or multiple pages of farmOS records"""
+        if params is None:
+            params = {}
 
-        logger.debug(
-            "Getting single record data for id: %s of entity type: %s",
-            id,
-            self.entity_type,
-        )
-        response = self.session.http_request(path=path)
+        params = {**self.params, **params}
 
+        path = self._get_resource_path(entity_type, bundle, resource_id)
+
+        response = self.session.http_request(path=path, params=params)
         return response.json()
 
-    def _get_record_data(self, filters):
-        """Retrieve one page of raw record data from the farmOS API."""
-        # Set path to return record type + filters
-        path = self.entity_type + ".json"
-        # Combine instance filters and filters from the method call
-        filters = {**self.filters, **filters}
+    def get(self, entity_type, bundle=None, params=None):
+        return self._get_records(entity_type=entity_type, bundle=bundle, params=params)
 
-        response = self.session.http_request(path=path, params=filters)
-
-        # Return object
-        data = {}
-
-        response = response.json()
-        if "list" in response:
-            data["list"] = response["list"]
-            data["page"] = {
-                "self": _parse_api_page(url=response["self"]),
-                "first": _parse_api_page(url=response["first"]),
-                "last": _parse_api_page(url=response["last"]),
-            }
-
-        return data
-
-    def _get_all_record_data(self, filters, page=0, list=None):
-        """Recursive function to retrieve multiple pages of raw record data from the farmOS API."""
-        if list is None:
-            list = []
-
-        filters["page"] = page
-
-        logger.debug(
-            "Getting page: %s of record data of entity type: %s",
-            filters["page"],
-            self.entity_type,
+    def get_id(self, entity_type, bundle=None, resource_id=None, params=None):
+        return self._get_records(
+            entity_type=entity_type,
+            bundle=bundle,
+            params=params,
+            resource_id=resource_id,
         )
-        data = self._get_record_data(filters=filters)
 
-        # Append record data to list of all requested data
-        if "list" in data:
-            list = list + data["list"]
+    def iterate(self, entity_type, bundle=None, params=None):
+        response = self._get_records(
+            entity_type=entity_type, bundle=bundle, params=params
+        )
+        more = True
+        while more:
+            # TODO: Should we merge in the "includes" info here?
+            yield from response["data"]
+            try:
+                next_url = response["links"]["next"]["href"]
+                parsed_url = urlparse(next_url)
+                next_path = parsed_url._replace(scheme="", netloc="").geturl()
+                response = self.session.http_request(path=next_path)
+                response = response.json()
+            except KeyError:
+                more = False
 
-        # Check to see if there are more pages
-        if "page" in data:
-            last_page = data["page"]["last"]
-            # Last page, return the list
-            if last_page == page:
-                data = {
-                    "page": {
-                        "first": data["page"]["first"],
-                        "last": data["page"]["last"],
-                    },
-                    "list": list,
-                }
-                return data
-            # Recursive call, get the next page
-            else:
-                return self._get_all_record_data(
-                    list=list, page=(page + 1), filters=filters
-                )
+    def send(self, entity_type, bundle=None, payload=None):
+        # Default to empty payload dict.
+        if payload is None:
+            payload = {}
 
-    def _get_records(self, filters=None):
-        """Helper function that checks to retrieve one record, one page or multiple pages of farmOS records"""
-        if filters is None:
-            filters = {}
-
-        # Determine if filters is an int (id) or dict (filters object)
-        if isinstance(filters, int) or isinstance(filters, str):
-            return self._get_single_record_data(filters)
-        elif isinstance(filters, dict):
-            # Check if the caller requests a specific page
-            if "page" in filters:
-                logger.debug(
-                    "Getting page: %s of record data of entity type: %s",
-                    filters["page"],
-                    self.entity_type,
-                )
-                return self._get_record_data(filters=filters)
-            else:
-                logger.debug(
-                    "Getting all record data of entity type: %s", self.entity_type
-                )
-                return self._get_all_record_data(filters=filters)
-
-    def get(self, filters=None):
-        """Simple get method"""
-
-        data = self._get_records(filters=filters)
-
-        return data
-
-    def send(self, payload):
-        options = {"json": payload}
+        # Set the resource type.
+        payload["type"] = self._get_resource_type(entity_type, bundle)
+        json_payload = {
+            "data": {**payload},
+        }
+        options = {"json": json_payload}
 
         # If an ID is included, update the record
         id = payload.pop("id", None)
         if id:
-            logger.debug("Updating record id: of entity type: %s", id, self.entity_type)
-            path = self.entity_type + "/" + str(id)
+            json_payload["data"]["id"] = id
+            logger.debug("Updating record id: of entity type: %s", id, entity_type)
+            path = self._get_resource_path(
+                entity_type=entity_type, bundle=bundle, record_id=id
+            )
             response = self.session.http_request(
-                method="PUT", path=path, options=options
+                method="PATCH", path=path, options=options
             )
         # If no ID is included, create a new record
         else:
-            logger.debug("Creating record of entity type: %s", self.entity_type)
-            path = self.entity_type
+            logger.debug("Creating record of entity type: %s", entity_type)
+            path = self._get_resource_path(entity_type=entity_type, bundle=bundle)
             response = self.session.http_request(
                 method="POST", path=path, options=options
             )
@@ -148,112 +86,141 @@ class BaseAPI:
         # Handle response from POST requests
         if response.status_code == 201:
             logger.debug("Record created.")
-            return response.json()
 
         # Handle response from PUT requests
         if response.status_code == 200:
             logger.debug("Record updated.")
-            # farmOS returns no response data for PUT requests
-            # response_data = response.json()
 
-            # Hard code the entity ID, path, and resource
-            entity_data = {
-                "id": id,
-                "uri": path,
-                "resource": self.entity_type,
-            }
-            return entity_data
+        return response.json()
 
-    def delete(self, id):
-        logger.debug("Deleted record id: %s of entity type: %s", id, self.entity_type)
-        path = self.entity_type + "/" + str(id)
-        response = self.session.http_request(method="DELETE", path=path)
+    def delete(self, entity_type, bundle=None, id=None):
+        logger.debug("Deleted record id: %s of entity type: %s", id, entity_type)
+        path = self._get_resource_path(
+            entity_type=entity_type, bundle=bundle, record_id=id
+        )
+        return self.session.http_request(method="DELETE", path=path)
 
-        return response
+    @staticmethod
+    def _get_resource_path(entity_type, bundle=None, record_id=None):
+        """Helper function that builds paths to jsonapi resources."""
+
+        if bundle is None:
+            bundle = entity_type
+
+        path = "api/" + entity_type + "/" + bundle
+
+        if record_id:
+            path += "/" + str(record_id)
+
+        return path
+
+    @staticmethod
+    def _get_resource_type(entity_type, bundle=None):
+        """Helper function that builds a JSONAPI resource name."""
+
+        if bundle is None:
+            bundle = entity_type
+
+        return entity_type + "--" + bundle
 
 
-class TermAPI(BaseAPI):
+class ResourceHelperBase:
+    def __init__(self, session, entity_type):
+        self.entity_type = entity_type
+        self.resource_api = ResourceBase(session=session)
+
+    def get(self, bundle, params=None):
+        return self.resource_api.get(
+            entity_type=self.entity_type, bundle=bundle, params=params
+        )
+
+    def get_id(self, bundle, resource_id, params=None):
+        return self.resource_api.get_id(
+            entity_type=self.entity_type,
+            bundle=bundle,
+            resource_id=resource_id,
+            params=params,
+        )
+
+    def iterate(self, bundle, params=None):
+        return self.resource_api.iterate(
+            entity_type=self.entity_type, bundle=bundle, params=params
+        )
+
+    def send(self, bundle, payload=None):
+        return self.resource_api.send(
+            entity_type=self.entity_type, bundle=bundle, payload=payload
+        )
+
+    def delete(self, bundle, id):
+        return self.resource_api.delete(
+            entity_type=self.entity_type, bundle=bundle, id=id
+        )
+
+
+class AssetAPI(ResourceHelperBase):
+    """API for interacting with farm assets"""
+
+    def __init__(self, session):
+        # Define 'asset' as the JSONAPI resource type.
+        super().__init__(session=session, entity_type="asset")
+
+
+class LogAPI(ResourceHelperBase):
+    """API for interacting with farm logs"""
+
+    def __init__(self, session):
+        # Define 'log' as the JSONAPI resource type.
+        super().__init__(session=session, entity_type="log")
+
+
+class TermAPI(ResourceHelperBase):
     """API for interacting with farm Terms"""
 
     def __init__(self, session):
         # Define 'taxonomy_term' as the farmOS API entity endpoint
         super().__init__(session=session, entity_type="taxonomy_term")
 
-    def get(self, filters=None):
-        """Get method that supports a bundle name as the 'filter' parameter"""
-
-        # Check if filters parameter is a str
-        if isinstance(filters, str):
-            # Add filters to instance requests.session filter dict with keyword 'bundle'
-            self.filters["bundle"] = filters
-            # Reset filters to empty dict
-            filters = {}
-
-        data = self._get_records(filters=filters)
-
-        return data
-
-
-class LogAPI(BaseAPI):
-    """API for interacting with farm logs"""
-
-    def __init__(self, session):
-        # Define 'log' as the farmOS API entity endpoint
-        super().__init__(session=session, entity_type="log")
-
-
-class AssetAPI(BaseAPI):
-    """API for interacting with farm assets"""
-
-    def __init__(self, session):
-        # Define 'farm_asset' as the farmOS API entity endpoint
-        super().__init__(session=session, entity_type="farm_asset")
-
-
-class AreaAPI(TermAPI):
-    """API for interacting with farm areas, a subset of farm terms"""
-
-    def __init__(self, session):
-        super().__init__(session=session)
-        self.filters["bundle"] = "farm_areas"
-
-    def get(self, filters=None):
-        """Retrieve raw record data from the farmOS API.
-
-        Override get() from BaseAPI to support TID (Taxonomy ID)
-        rather than record ID
-        """
-
-        # Determine if filters is an int (tid) or dict (filters object)
-        if isinstance(filters, int) or isinstance(filters, str):
-            tid = str(filters)
-            # Add tid to filters object
-            filters = {
-                "tid": tid,
-            }
-
-        data = self._get_records(filters=filters)
-
-        return data
-
 
 def info(session):
     """Retrieve info about the farmOS server."""
 
     logger.debug("Retrieving farmOS server info.")
-    response = session.http_request("farm.json")
+    response = session.http_request("api")
     return response.json()
 
 
-def _parse_api_page(url):
-    """Helper function that returns page numbers from the API response.
+def filter(path, value, operation="="):
+    """Helper method to build JSONAPI filter query params."""
 
-    The farmOS API returns URLs for the self, first and last pages of records
-    in requests that return many records. This function helps parse a raw page
-    number from the returned URL.
-    """
+    # TODO: Validate path, value, operation?
+    # TODO: Support filter groups.
+    # TODO: Support revision filters, are these edge cases?
+    # TODO: Support "pagination" query params? Separate method?
+    # TODO: Support "sort" query params? Separate method?
+    # TODO: Support "includes" query params? Separate method?
 
-    parsed_url = urlparse(url)
-    page_num = parse_qs(parsed_url.query)["page"][0]
+    filters = {}
 
-    return int(page_num)
+    # If the operation is '=', only one query param is required.
+    if operation == "=":
+        param = f"filter[{path}]"
+        filters[param] = value
+    # Else we need a query param for the path, operation, and value.
+    else:
+        # TODO: Use a unique identifier instead of using "condition"
+        # otherwise the same path cannot be filtered multiple times.
+        base_param = f"filter[{path}_{operation.lower()}][condition]"
+
+        path_param = base_param + "[path]"
+        filters[path_param] = path
+
+        op_param = base_param + "[operator]"
+        filters[op_param] = operation
+
+        value_param = base_param + "[value]"
+        if operation in ["IN", "NOT IN", ">", "<", "<>", "BETWEEN"]:
+            value_param += "[]"
+        filters[value_param] = value
+
+    return filters
